+++ examples/with-ethers/src/createNewEthereumPrivateKey.ts
@@ -46,12 +46,22 @@ export async function createNewEthereumPrivateKey() {
       activity.result.createPrivateKeysResult?.privateKeyIds?.[0]
     );
 
+    const keyInfo = await TurnkeyApi.postGetPrivateKey({
+      body: {
+        organizationId: process.env.ORGANIZATION_ID!,
+        privateKeyId,
+      },
+    });
+
+    const address = refineNonNull(keyInfo.privateKey.addresses[0]?.address);
+
     // Success!
     console.log(
       [
         `New Ethereum private key created!`,
         `- Name: ${privateKeyName}`,
         `- Private key ID: ${privateKeyId}`,
+        `- Address: ${address}`,
         ``,
         "Now you can take the private key ID, put it in `.env.local`, then re-run the script.",
       ].join("\n")