+++ packages/http/src/index.ts
@@ -2,7 +2,7 @@ import { PublicApiService as TurnkeyApi } from "./__generated__/barrel";
 
 export { init } from "./config";
 
-export { TurnkeyActivityError } from "./shared";
+export { TurnkeyActivityError, TurnkeyRequestError } from "./shared";
 
 export { withAsyncPolling } from "./async";
 
@@ -14,3 +14,5 @@ export { TurnkeyApi };
 const PublicApiService = TurnkeyApi;
 
 export { PublicApiService };
+
+export { sealAndStampRequestBody } from "./base";