+++ packages/http/src/shared.ts
@@ -39,3 +39,28 @@ export type TStamper = (input: {
   scheme: string;
   signature: string;
 }>;
+
+export type GrpcStatus = {
+  message: string;
+  code: number;
+  details: unknown[] | null;
+};
+
+export class TurnkeyRequestError extends Error {
+  details: any[] | null;
+  code: number;
+
+  constructor(input: GrpcStatus) {
+    let turnkeyErrorMessage = `Turnkey error ${input.code}: ${input.message}`;
+
+    if (input.details != null) {
+      turnkeyErrorMessage += ` (Details: ${JSON.stringify(input.details)})`;
+    }
+
+    super(turnkeyErrorMessage);
+
+    this.name = "TurnkeyRequestError";
+    this.details = input.details ?? null;
+    this.code = input.code;
+  }
+}