+++ packages/ethers/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @turnkey/ethers
 
+## 0.15.0
+
+### Minor Changes
+
+- No public facing changes
+
+### Patch Changes
+
+- Updated dependencies
+  - @turnkey/http@0.16.0
+    - Fix `.postGetPrivateKey(...)`'s underlying path, while adding `@deprecated` `.postGetPrivateKeyBackwardsCompat(...)` for backward compatibility
+
+## 0.14.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @turnkey/http@0.15.0
+
+## 0.14.0
+
+### Minor Changes
+
+- `signTransaction(...)` now verifies and drops `tx.from` if present
+  - This mimics the behavior of ethers' Wallet [implementation](https://github.com/ethers-io/ethers.js/blob/f97b92bbb1bde22fcc44100af78d7f31602863ab/packages/wallet/src.ts/index.ts#L117-L121)
+
+### Patch Changes
+
+- Updated dependencies
+  - @turnkey/http@0.14.0
+
+## 0.13.2
+
+### Patch Changes
+
+- New `TurnkeyRequestError` error class that contains rich error details
+- Updated dependencies
+  - @turnkey/http@0.13.2
+
+## 0.13.1
+
+### Patch Changes
+
+- Error messages now contain Turnkey-specific error details
+- Updated dependencies
+  - @turnkey/http@0.13.1
+
 ## 0.13.0
 
 ### Minor Changes